Wrap canvas in a Renderer
Rather than calling draw methods directly on the canvas, this wraps it
in a Renderer module. The benefit of this approach is that when we
want to do more advanced rendering with images and fonts, we can keep
the drawing API simple and hide the texture/font loading in the
Renderer.

This change just replaces the existing usage on canvas with equivalent
methods on the Renderer. A subsequent change will introduce textures.