Render textured bricks (#9)
This renders bricks as a texture rather than a filled rect. The
renderer handles loading the assets and mapping the requested image to
the appropriate spot in the texture.

It also renders the ghost piece at half opacity by setting the alpha
channel on the texture.

The `image` feature in `sdl2` enables the SDL2_image extension which
enables loading files from PNG. You'll need to install `sdl2_image`
via Homebrew for this extension to work.

The `unsafe_textures` feature eliminates some [annoying lifetimes][0]
for textures. Since the texture will live until the end of the
process, they're basically static and this isn't really a big deal.

Right now everything is rendered as a red brick. I'll add support for
other brick types in a subsequent PR.

[0]: https://github.com/Rust-SDL2/rust-sdl2#about-the-unsafe_textures-feature