Parse host and port from args
This adds command-line parsing to both the client and server so we can
specify which IP and port to connect to. This will make it easier to
test it out on an actual server without having to change the address
in the code.

To use with `cargo run` you'll need to add a `--` before passing in
the flags.

```
$ cargo run --bin server -- --port 4489
$ cargo run --bin client -- --port 4489 --host 10.0.0.0
```