Add a connect scene after title screen
The connect scene will connect to the server and fetch the initial
state. Handling this in a separate scene rather than the setup code
will make it easier to switch to non-blocking sockets since we'll be
able to use the game loop to continuously poll the server for the game
state.

Since we're not connecting right away on startup, the boot order of
the client and server doesn't matter much. This makes it easier to use
something like `foreman` to startup both processes. I added a
`Procfile` that will launch the client and server, although the server
exits right away so it kills both processes once the game starts. I'll
address this in a followup PR.