Wrap grid sync in a smart pointer (Cow)
When the server serializes the grid state to the client, we don't want
to move ownership of the grid into the serialized message. But when
the client receives the new grid state, it should be owned because
we're creating it.

Because we're sharing the same message struct on the client and the
server, we want to be able to serialize it using a reference and
deserialize it into an owned value. The [copy-on-write (Cow)][1] smart
pointer lets us do just this with a uniform interface. I got the idea
from [this][0] Stackoverflow post.

[0]: https://stackoverflow.com/a/52733564/181403
[1]: https://doc.rust-lang.org/std/borrow/enum.Cow.html