Use non-blocking sockets
Since we don't want to pause the client or server when there are no
messages queued up, we set the nonblocking flag on the socket. We
change the interface to return a `Result<Option<(SocketAddr, Message)>>`
with an `Ok(None)` value indicating that there were no messages
available.

This also returns `Ok(None)` if we receive a message we don't know how
to deserialize. We might want to log this information, although it's
not clear what we'd want to log and even printing out a line would be
expensive if we're receiving a bunch of unwanted packets. We can
revisit this later when we make the network layer more robust.

This code pulls from the [nonblocking example in the
documentation][0].

[0]: https://doc.rust-lang.org/std/net/struct.UdpSocket.html#method.set_nonblocking