Use non-blocking sockets (#26)
* Wrap grid sync in a smart pointer (Cow)

When the server serializes the grid state to the client, we don't want
to move ownership of the grid into the serialized message. But when
the client receives the new grid state, it should be owned because
we're creating it.

Because we're sharing the same message struct on the client and the
server, we want to be able to serialize it using a reference and
deserialize it into an owned value. The [copy-on-write (Cow)][1] smart
pointer lets us do just this with a uniform interface. I got the idea
from [this][0] Stackoverflow post.

[0]: https://stackoverflow.com/a/52733564/181403
[1]: https://doc.rust-lang.org/std/borrow/enum.Cow.html

* Use non-blocking sockets

Since we don't want to pause the client or server when there are no
messages queued up, we set the nonblocking flag on the socket. We
change the interface to return a `Result<Option<(SocketAddr, Message)>>`
with an `Ok(None)` value indicating that there were no messages
available.

This also returns `Ok(None)` if we receive a message we don't know how
to deserialize. We might want to log this information, although it's
not clear what we'd want to log and even printing out a line would be
expensive if we're receiving a bunch of unwanted packets. We can
revisit this later when we make the network layer more robust.

This code pulls from the [nonblocking example in the
documentation][0].

[0]: https://doc.rust-lang.org/std/net/struct.UdpSocket.html#method.set_nonblocking