Use canvas logical size to handle window scaling (#72)
SDL supports setting a logical size on a canvas which is independent
of the actual window size. Everything in the code is positioned via
the logical size, and as the window size changes SDL handles scaling
to the appropriate size as necessary. If the aspect ratio differs
between the logical size and the window size, it adds blank lines on
the top and bottom or sides.

This handles the issue where the title screen was not resizing when
the window size changed.

Fixes #57