Adds new member to Scene trait to allow us to quit the game cleanly from various locations (#74)
The way our API was set up there wasn't really a clean way for some sort of action deep within a scene to trigger a quitting of the game. This PR adds a new member to the Scene trait which enables a scene to set a flag that will break out of the game loop.

I'm using this in the title menu scene but I could see in the future also having a game menu that is accessible via 'escape' or some other key and wanting to quit from there.