Move socket ownership to game loop
The socket used for sending and receiving messages is created during
the connect scene and then passed around to subsequent scenes. Each
scene was responsible for polling for new messages on the socket.

I'd like to move the polling for new messages out of the scenes and
into the main game loop so we can ensure the messages are getting
processed regularly and standardize how network errors are handled.

To do this, we'll need a socket to poll within the game loop. This
change moves the socket out of the individual scenes and into the game
loop. The socket is then explicitly passed into each scene in case
they need to send or receive messages.

In a subsequent PR, I'll move polling for messages to the game loop so
that the scenes only need to handle messages.