Remove message handling from scene update
This builds on top of #83. Each scene was responsible for polling for
messages on the network and handling errors. Each scene did it
slightly differently, and some scenes didn't poll at all which means
messages would get backed up on the socket.

Also, the scenes only fetched one message from the socket each frame,
so if the client was receiving messages faster than the frame rate, it
would cause a backup of messages.

This change moves the polling the socket to the main game loop and
introduces a method to the Scene trait for handling those
messages. Even if the scene doesn't respond to any messages, this
ensures that we're actively processing the messages so that they don't
backup.

This also means we can standardize how we handle network errors or
invalid messages.