Reset rendering offsets automatically
Dealing with rendering offsets is painful because you have to:
* remember to reset them after use
* compute the absolute position based off existing offsets

This change replaces the `get_offset / set_offset` functions with a
single function that takes a closure to execute with the given
offsets. When the function returns, the offsets will have been reset
to their original value.

The offsets specified are relative to the current offset, so this can
be used recursively for nested offsets.